Moves sets distribution in stats category
