+++ lib/pages/team-details/ranking/evolution_widget.dart
@@ -36,7 +36,7 @@ class EvolutionWidget extends StatelessWidget {
   @override
   Widget build(BuildContext context) {
     return Container(
-      padding: EdgeInsets.symmetric(vertical: 15.0),
+      padding: EdgeInsets.symmetric(vertical: 20.0),
       child: Row(
         children: <Widget>[
           Expanded(flex: 1, child: Text(title, textAlign: TextAlign.center, style: Theme.of(context).textTheme.subtitle2)),
